Move PulsarWorkerOpenTelemetry to org.apache.pulsar.functions.worker
